requirements.txt
openai>=1.0.0
openai-whisper
pyobjc
sounddevice
numpy
SpeechRecognition
pyaudio
pocketsphinx
pynput>=1.7.6

